% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bolBands.R
\name{bolBands}
\alias{bolBands}
\title{Create and plot Bollinger Bands}
\usage{
bolBands(data, n, dev, filtSeries)
}
\arguments{
\item{data}{A long dataframe containing 'date', 'series', and 'price'}

\item{n}{The number of periods for Exponential Moving Average}

\item{dev}{Standard deviation of trend line}

\item{filtSeries}{Filter to apply to series (default: First series in data)}
}
\value{
A list containing the following:

\item{data}{dataframe containing Bollinger Band requirements `tibble`}
\item{plot}{A plot of the Bollinger Bands with the n-day SMA}
}
\description{
Takes a dataframe with price, creates Bollinger Bands,
plots an interactive chart and returns a dataframe
}
\examples{
bolBands(data = topTech, n = 20, dev = 2)
bolBands(data = topTech, n = 20, dev = 2, filtSeries = "AAPL")

}
