% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{indexes}
\alias{indexes}
\title{dataset: IWM, and FXI pricing for last year}
\format{
data frame
}
\source{
{Yahoo Finance}
}
\usage{
indexes
}
\value{
`tibble`
}
\description{
Adjusted stock prices for IWM and FXI
}
\keyword{datasets}
