% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fluf.R
\name{plot_fluf}
\alias{plot_fluf}
\title{Create an interactive plot using Plotly from any Yahoo Finance ticker. This function is designed to give the user a quick glanc at the last year for the stock.}
\usage{
plot_fluf(ticker)
}
\arguments{
\item{ticker}{A string of a Yahoo Finance ticker}
}
\value{
An interactive Plotly plot.
}
\description{
This function takes a Yahoo Finance Ticker as input and generates an interactive line plot
using the Plotly library.
}
\examples{
plot_fluf("TSLA")

}
