% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotter.R
\name{plotter}
\alias{plotter}
\title{Create an interactive plot using Plotly from a long dataframe.}
\usage{
plotter(data)
}
\arguments{
\item{data}{A long dataframe with columns 'date', 'series', and 'value'.}
}
\value{
An interactive Plotly plot.
}
\description{
This function takes a long dataframe as input and generates an interactive line plot
using the Plotly library.
}
\examples{
plotter(snp)

}
