% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleregression.R
\name{simpleregression}
\alias{simpleregression}
\title{Runs a simple linear regression on two variables}
\usage{
simpleregression(data)
}
\arguments{
\item{data}{A long dataframe with columns 'dependent' and 'independent'.}
}
\value{
Linear regression stats
}
\description{
This function takes two variables and runs a simple linear regression and outputs the stats
}
\examples{
simpleregression(data)

}
