% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{topTech}
\alias{topTech}
\title{dataset: Apple, Microsoft, and Amazon pricing for last year}
\format{
data frame
}
\source{
{Yahoo Finance}
}
\usage{
topTech
}
\value{
`tibble`
}
\description{
Adjusted stock prices for Apple, Microsoft, and Amazon
}
\keyword{datasets}
