% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleregression.R
\name{simpleregression}
\alias{simpleregression}
\title{Runs a simple linear regression on two specified variables in a dataframe}
\usage{
simpleregression(data, indep_col, dep_col)
}
\arguments{
\item{data}{A dataframe containing at least two columns for regression analysis.}

\item{indep_col}{Name of the independent variable column.}

\item{dep_col}{Name of the dependent variable column.}
}
\value{
Linear regression stats
}
\description{
This function takes a dataframe and two column names, then runs a simple linear regression and outputs the stats
}
\examples{
data <- data.frame(MSFT = c(2, 4, 6, 8), AAPL = c(10, 12, 14, 20))
simpleregression(data, 'MSFT', 'AAPL')

}
